fix: don't build FUSE paths for Windows
Fixes #1390.
